feat: add first two action templates
These will be deployed once for the dao, have to be registered as extensions, and can be disabled the same way as other extensions. Any action contract created this way can be deployed then submitted as an action to the action-proposal contract. If the vote succeeds the run() function is performed. This way it runs the code in the action versus arbitrary clarity - and that code can be simple like this or as complex as needed.