BREAKING: rearrange tools into correct folders
The generate/deploy patterns are now used for creating the dao at the top level, within the core-proposals folder at the top level, and within the smart-wallet folder at the top level for consistency. The file naming wasn't lining up and changing the order would make it more confusing. Also moves the construct-dao tool to the correct location as it's a public function call to the base-dao contract.