fix: update script to use simplified clarity type
There was a mismatch between what Cl.principal() creates in Stacks 6.x and 7.x, creating an odd error on parsing the values even though they should be valid. Long-term fix will be to accept serialized values or find a way to better convert between the formats. Using the simple workaround for now.